Game-2 screen almost done
