--- js/modules/templates/stats/handler/stats-view-handler.js
@@ -1,1 +1,0 @@
-
